rename ressource_uri package to com.celements.filebase.uri
