+++ src/main/java/com/celements/filebase/uri/UrlRessourceNotExistException.java
@@ -1,4 +1,4 @@
-package com.celements.ressource_url;
+package com.celements.filebase.uri;
 
 import javax.validation.constraints.NotNull;
